refactor controller in react context
